untested constraints for schedule

+++ Development/org.mdse.pts.schedule/src-gen/org/mdse/pts/schedule/Time.java
@@ -27,12 +27,12 @@ public interface Time extends EObject {
 	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
 	 * @return the value of the '<em>Hour</em>' attribute.
-	 * @see #setHour(byte)
+	 * @see #setHour(int)
 	 * @see org.mdse.pts.schedule.SchedulePackage#getTime_Hour()
 	 * @model required="true"
 	 * @generated
 	 */
-	byte getHour();
+	int getHour();
 
 	/**
 	 * Sets the value of the '{@link org.mdse.pts.schedule.Time#getHour <em>Hour</em>}' attribute.
@@ -42,19 +42,19 @@ public interface Time extends EObject {
 	 * @see #getHour()
 	 * @generated
 	 */
-	void setHour(byte value);
+	void setHour(int value);
 
 	/**
 	 * Returns the value of the '<em><b>Minute</b></em>' attribute.
 	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
 	 * @return the value of the '<em>Minute</em>' attribute.
-	 * @see #setMinute(byte)
+	 * @see #setMinute(int)
 	 * @see org.mdse.pts.schedule.SchedulePackage#getTime_Minute()
 	 * @model required="true"
 	 * @generated
 	 */
-	byte getMinute();
+	int getMinute();
 
 	/**
 	 * Sets the value of the '{@link org.mdse.pts.schedule.Time#getMinute <em>Minute</em>}' attribute.
@@ -64,6 +64,6 @@ public interface Time extends EObject {
 	 * @see #getMinute()
 	 * @generated
 	 */
-	void setMinute(byte value);
+	void setMinute(int value);
 
 } // Time
